+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 # CHANGES
 
+## 2021-07-12
+
+* Add script to plot the figures in AHM2 section 8.2.3 (Figs 8.4 to 8.7) with credible intervals.
+
 ## 2021-06-25
 
 * Incorporate [errata](https://sites.google.com/site/appliedhierarchicalmodeling/errata) into code for sections AHM2 3.4.4 and 3.4.5.